Move and rename lutron caseta base entity to separate module (#126103)
