+++ homeassistant/components/lutron_caseta/fan.py
@@ -18,7 +18,7 @@ from homeassistant.util.percentage import (
     percentage_to_ordered_list_item,
 )
 
-from . import LutronCasetaDeviceUpdatableEntity
+from .entity import LutronCasetaUpdatableEntity
 from .models import LutronCasetaConfigEntry
 
 DEFAULT_ON_PERCENTAGE = 50
@@ -41,7 +41,7 @@ async def async_setup_entry(
     async_add_entities(LutronCasetaFan(fan_device, data) for fan_device in fan_devices)
 
 
-class LutronCasetaFan(LutronCasetaDeviceUpdatableEntity, FanEntity):
+class LutronCasetaFan(LutronCasetaUpdatableEntity, FanEntity):
     """Representation of a Lutron Caseta fan. Including Fan Speed."""
 
     _attr_supported_features = (
